requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
pymysql==1.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
